# docs/architecture/openapi.yaml
openapi: 3.1.0
info:
  title: SIYU API
  version: 0.1.0
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  description: |
    四时有余核心 REST API 契约。金额使用整数分且不得超过 JavaScript 安全整数；
    业务日期使用 YYYY-MM-DD，系统时间使用 UTC ISO 8601。
servers:
  - url: /api/v1
...
security:
  - bearerAuth: []
paths:
...
  /categories:
    get:
      tags: [Categories]
      operationId: listCategories
      summary: 获取当前成员可见的账本分类与服务端权限能力
      parameters:
        - $ref: '#/components/parameters/LedgerId'
        - in: query
          name: type
          schema: { $ref: '#/components/schemas/EntryType' }
        - in: query
          name: includeDisabled
          schema: { type: boolean, default: false }
      responses:
        '200': { $ref: '#/components/responses/CategoryListOk' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      tags: [Categories]
      operationId: createCategory
      summary: 在账本中幂等创建自定义分类
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateCategoryRequest' }
      responses:
        '201': { $ref: '#/components/responses/ResourceCreated' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '429': { $ref: '#/components/responses/RateLimited' }
...
  /statistics/overview:
    get:
      tags: [Statistics]
      operationId: getStatisticsOverview
      summary: 获取账本统计概览
      parameters:
        - $ref: '#/components/parameters/LedgerId'
        - $ref: '#/components/parameters/StatisticsMonth'
      responses:
        '200': { $ref: '#/components/responses/StatisticsOverviewOk' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /statistics/trend:
    get:
      tags: [Statistics]
      operationId: getStatisticsTrend
      summary: 获取账本收支趋势
      parameters:
        - $ref: '#/components/parameters/LedgerId'
        - $ref: '#/components/parameters/StatisticsMonth'
      responses:
        '200': { $ref: '#/components/responses/StatisticsTrendOk' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /statistics/categories:
    get:
      tags: [Statistics]
      operationId: getStatisticsCategories
      summary: 获取账本分类统计
      parameters:
        - $ref: '#/components/parameters/LedgerId'
        - $ref: '#/components/parameters/StatisticsMonth'
      responses:
        '200': { $ref: '#/components/responses/StatisticsCategoriesOk' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /statistics/members:
    get:
      tags: [Statistics]
      operationId: getStatisticsMembers
      summary: 获取情侣成员对比
      parameters:
        - $ref: '#/components/parameters/LedgerId'
        - $ref: '#/components/parameters/StatisticsMonth'
      responses:
        '200': { $ref: '#/components/responses/StatisticsMembersOk' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshCookie:
      type: apiKey
      in: cookie
      name: siyu_refresh_token
  parameters:
...
    LedgerId:
      in: query
      name: ledgerId
      required: true
      schema: { type: string, format: uuid }
...
    StatisticsMonth:
      in: query
      name: month
      description: 缺省为用户时区当前月
      schema: { type: string, pattern: '^\d{4}-(0[1-9]|1[0-2])$' }
...
  responses:
...
    ResourceCreated:
      description: 创建成功
      headers:
        X-Request-ID: { $ref: '#/components/headers/RequestId' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ResourceResponse' }
...
    CategoryListOk:
      description: 分类列表与服务端权限能力查询成功
      headers:
        X-Request-ID: { $ref: '#/components/headers/RequestId' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/CategoryListResponse' }
...
    StatisticsOverviewOk:
      description: 账本月度统计概览查询成功
      content:
        application/json:
          schema: { $ref: '#/components/schemas/StatisticsOverviewResponse' }
    StatisticsTrendOk:
      description: 账本完整逐日收支趋势查询成功
      content:
        application/json:
          schema: { $ref: '#/components/schemas/StatisticsTrendResponse' }
    StatisticsCategoriesOk:
      description: 账本支出分类统计查询成功
      content:
        application/json:
          schema: { $ref: '#/components/schemas/StatisticsCategoriesResponse' }
    StatisticsMembersOk:
      description: 账本成员支出统计查询成功
      content:
        application/json:
          schema: { $ref: '#/components/schemas/StatisticsMembersResponse' }
...
    ValidationFailed:
      description: 请求参数不合法
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unauthorized:
      description: 未认证或会话已失效
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: 无权执行目标操作
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: 资源不存在或按安全策略不可见
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Conflict:
      description: 幂等、唯一性、并发或业务状态冲突
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    RateLimited:
      description: 请求过于频繁
      headers:
        Retry-After:
          schema: { type: integer, minimum: 1 }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  headers:
    RequestId:
      description: 当前请求追踪标识
      schema: { type: string, minLength: 8, maxLength: 128 }
  schemas:
    Cent:
      type: integer
      format: int64
      minimum: 1
      maximum: 9007199254740991
      description: 整数分，不得超过 JavaScript 安全整数。
    NonNegativeCent:
      type: integer
      format: int64
      minimum: 0
      maximum: 9007199254740991
    BusinessDate:
      type: string
      format: date
    Timestamp:
      type: string
      format: date-time
      description: UTC ISO 8601 时间。
    IdempotencyKey:
      type: string
      minLength: 8
      maxLength: 128
      pattern: '^[A-Za-z0-9._:-]+$'
    EntryType:
      type: string
      enum: [INCOME, EXPENSE]
    EntrySourceType:
      type: string
      enum: [MANUAL, SALARY, DEBT_TRANSACTION, RECURRING_RUN]
    EntryPaymentMethod:
      type: string
      enum: [CASH, WECHAT, ALIPAY, BANK_CARD, OTHER]
    CategoryIconKey:
      type: string
      enum:
        [food, shopping, transport, housing, entertainment, medical, education, gift, salary, bonus, part_time, investment, red_packet, refund, other]
    CategoryColor:
      type: string
      pattern: '^#[0-9A-Fa-f]{6}$'
    RecurringRunStatus:
      type: string
      enum: [PENDING, GENERATED, CONFIRMED, SKIPPED, FAILED]
    User:
      type: object
      required: [id, nickname, timezone, status, email, roles, permissions]
      properties:
        id: { type: string, format: uuid }
        nickname: { type: string, minLength: 1, maxLength: 100 }
        avatarUrl: { type: [string, 'null'], format: uri }
        timezone: { type: string, minLength: 1, maxLength: 64 }
        status: { type: string, enum: [ACTIVE, DISABLED] }
        email: { type: [string, 'null'], format: email }
        roles:
          type: array
          uniqueItems: true
          items: { type: string, enum: [USER, ADMIN] }
        permissions:
          type: array
          uniqueItems: true
          items: { type: string }
    Ledger:
      type: object
      required: [id, type, name, ownerUserId, status, members]
      properties:
        id: { type: string, format: uuid }
        type: { type: string, enum: [PERSONAL, COUPLE] }
        name: { type: string, minLength: 1, maxLength: 100 }
        ownerUserId: { type: string, format: uuid }
        status: { type: string, enum: [ACTIVE, DISSOLVED] }
        members:
          type: array
          maxItems: 2
          items: { $ref: '#/components/schemas/LedgerMember' }
    LedgerMember:
      type: object
      required: [userId, role, joinedAt, nickname]
      properties:
        userId: { type: string, format: uuid }
        role: { type: string, enum: [OWNER, MEMBER] }
        joinedAt: { $ref: '#/components/schemas/Timestamp' }
        nickname: { type: string, minLength: 1, maxLength: 100 }
        avatarUrl: { type: [string, 'null'], format: uri }
...
    Category:
      type: object
      required:
        [id, ledgerId, creatorUserId, type, name, icon, color, sortOrder, isSystem, isEnabled, canEdit, canToggle, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        ledgerId: { type: string, format: uuid }
        creatorUserId: { type: [string, 'null'], format: uuid }
        type: { $ref: '#/components/schemas/EntryType' }
        name: { type: string, minLength: 1, maxLength: 50 }
        icon: { $ref: '#/components/schemas/CategoryIconKey' }
        color: { $ref: '#/components/schemas/CategoryColor' }
        sortOrder: { type: integer, minimum: 0 }
        isSystem: { type: boolean }
        isEnabled: { type: boolean }
        canEdit: { type: boolean }
        canToggle: { type: boolean }
        createdAt: { $ref: '#/components/schemas/Timestamp' }
        updatedAt: { $ref: '#/components/schemas/Timestamp' }
    Entry:
      type: object
      required:
        [id, ledgerId, type, amountCent, businessDate, note, paymentMethod, sourceType, creator, category, createdAt, updatedAt, version, canEdit, canDelete]
      properties:
        id: { type: string, format: uuid }
        ledgerId: { type: string, format: uuid }
        type: { $ref: '#/components/schemas/EntryType' }
        amountCent: { $ref: '#/components/schemas/Cent' }
        businessDate: { $ref: '#/components/schemas/BusinessDate' }
        note: { type: [string, 'null'], maxLength: 500 }
        paymentMethod:
          anyOf:
            - $ref: '#/components/schemas/EntryPaymentMethod'
            - type: 'null'
        sourceType: { $ref: '#/components/schemas/EntrySourceType' }
        creator: { $ref: '#/components/schemas/EntryCreator' }
        category: { $ref: '#/components/schemas/EntryCategorySummary' }
        createdAt: { $ref: '#/components/schemas/Timestamp' }
        updatedAt: { $ref: '#/components/schemas/Timestamp' }
        version: { type: integer, minimum: 1 }
        canEdit: { type: boolean }
        canDelete: { type: boolean }
    EntryCreator:
      type: object
      required: [id, nickname, avatarUrl]
      properties:
        id: { type: string, format: uuid }
        nickname: { type: string, minLength: 1, maxLength: 100 }
        avatarUrl: { type: [string, 'null'], format: uri, maxLength: 2048 }
    EntryCategorySummary:
      type: object
      required: [id, name, icon, color, isEnabled]
      properties:
        id: { type: string, format: uuid }
        name: { type: string, minLength: 1, maxLength: 50 }
        icon: { $ref: '#/components/schemas/CategoryIconKey' }
        color: { $ref: '#/components/schemas/CategoryColor' }
        isEnabled: { type: boolean }
    Debt:
      type: object
      required: [id, direction, counterpartyName, principalCent, processedCent, remainingCent, startDate, dueDate, status, overdueDays, note, reminderEnabled, canEdit, canDelete, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        direction: { type: string, enum: [BORROWED, LENT] }
        counterpartyName: { type: string, minLength: 1, maxLength: 100 }
        principalCent: { $ref: '#/components/schemas/Cent' }
        processedCent: { $ref: '#/components/schemas/NonNegativeCent' }
        remainingCent: { $ref: '#/components/schemas/NonNegativeCent' }
        startDate: { $ref: '#/components/schemas/BusinessDate' }
        dueDate: { anyOf: [{ $ref: '#/components/schemas/BusinessDate' }, { type: 'null' }] }
        status: { type: string, enum: [ACTIVE, SETTLED, OVERDUE, CANCELLED] }
        overdueDays: { type: integer, minimum: 0 }
        note: { type: [string, 'null'], maxLength: 500 }
        reminderEnabled: { type: boolean }
        canEdit: { type: boolean, readOnly: true }
        canDelete: { type: boolean, readOnly: true }
        createdAt: { type: string, format: date-time, readOnly: true }
        updatedAt: { type: string, format: date-time, readOnly: true }
        transactions:
          type: array
          items: { $ref: '#/components/schemas/DebtTransaction' }
    DebtTransaction:
      type: object
      required: [id, debtId, userId, amountCent, businessDate, syncEntry, entryId, note, createdAt]
      properties:
        id: { type: string, format: uuid }
        debtId: { type: string, format: uuid }
        userId: { type: string, format: uuid }
        amountCent: { $ref: '#/components/schemas/Cent' }
        businessDate: { $ref: '#/components/schemas/BusinessDate' }
        syncEntry: { type: boolean }
        entryId: { type: [string, 'null'], format: uuid }
        note: { type: [string, 'null'], maxLength: 500 }
        createdAt: { type: string, format: date-time, readOnly: true }
    RecurringRule:
      type: object
      required: [id, ledgerId, name, entryType, amountCent, categoryId, frequency, generationMode, status]
      properties:
        id: { type: string, format: uuid }
        ledgerId: { type: string, format: uuid }
        name: { type: string, minLength: 1, maxLength: 100 }
        entryType: { $ref: '#/components/schemas/EntryType' }
        amountCent: { $ref: '#/components/schemas/Cent' }
        categoryId: { type: string, format: uuid }
        frequency: { type: string, enum: [MONTHLY, YEARLY] }
        generationMode: { type: string, enum: [AUTO, CONFIRM] }
        status: { type: string, enum: [ACTIVE, PAUSED, COMPLETED, CANCELLED] }
    RecurringRun:
      type: object
      required: [id, ruleId, scheduledDate, amountCent, status, attempts]
      properties:
        id: { type: string, format: uuid }
        ruleId: { type: string, format: uuid }
        scheduledDate: { $ref: '#/components/schemas/BusinessDate' }
        amountCent: { $ref: '#/components/schemas/Cent' }
        status: { $ref: '#/components/schemas/RecurringRunStatus' }
        entryId: { type: [string, 'null'], format: uuid }
        attempts: { type: integer, minimum: 0 }
        lastError: { type: [string, 'null'] }
    SalaryProfile:
      type: object
      required: [id, name, payDay, defaultSyncEntry]
      properties:
        id: { type: string, format: uuid }
        name: { type: string, minLength: 1, maxLength: 100 }
        employerName: { type: [string, 'null'], maxLength: 100 }
        payDay: { type: integer, minimum: 1, maximum: 31 }
        defaultSyncEntry: { type: boolean }
    SalaryRecord:
      type: object
      required: [id, profileId, salaryMonth, grossCent, deductionCent, netCent, paymentStatus, items]
      properties:
        id: { type: string, format: uuid }
        profileId: { type: string, format: uuid }
        salaryMonth: { $ref: '#/components/schemas/BusinessDate' }
        grossCent: { $ref: '#/components/schemas/NonNegativeCent' }
        deductionCent: { $ref: '#/components/schemas/NonNegativeCent' }
        netCent: { $ref: '#/components/schemas/NonNegativeCent' }
        paymentStatus: { type: string, enum: [UNPAID, PAID] }
        items:
          type: array
          items: { $ref: '#/components/schemas/SalaryItem' }
    SalaryItem:
      type: object
      required: [itemType, itemCode, itemName, amountCent]
      properties:
        itemType: { type: string, enum: [EARNING, DEDUCTION] }
        itemCode: { type: string, minLength: 1, maxLength: 50 }
        itemName: { type: string, minLength: 1, maxLength: 100 }
        amountCent: { $ref: '#/components/schemas/Cent' }
    SavingGoal:
      type: object
      required: [id, ledgerId, creatorUserId, name, targetCent, initialCent, savedCent, status]
      properties:
        id: { type: string, format: uuid }
        ledgerId: { type: string, format: uuid }
        creatorUserId: { type: string, format: uuid }
        name: { type: string, minLength: 1, maxLength: 100 }
        targetCent: { $ref: '#/components/schemas/Cent' }
        initialCent: { $ref: '#/components/schemas/NonNegativeCent' }
        savedCent: { $ref: '#/components/schemas/NonNegativeCent' }
        targetDate: { anyOf: [{ $ref: '#/components/schemas/BusinessDate' }, { type: 'null' }] }
        status: { type: string, enum: [ACTIVE, COMPLETED, CANCELLED] }
    SavingContribution:
      type: object
      required: [id, goalId, userId, amountCent, businessDate]
      properties:
        id: { type: string, format: uuid }
        goalId: { type: string, format: uuid }
        userId: { type: string, format: uuid }
        amountCent: { $ref: '#/components/schemas/Cent' }
        businessDate: { $ref: '#/components/schemas/BusinessDate' }
        note: { type: [string, 'null'], maxLength: 500 }
    Notification:
      type: object
      required: [id, type, title, content, createdAt]
      properties:
        id: { type: string, format: uuid }
        type: { type: string, minLength: 1, maxLength: 50 }
        title: { type: string, minLength: 1, maxLength: 100 }
        content: { type: string, minLength: 1, maxLength: 1000 }
        readAt: { anyOf: [{ $ref: '#/components/schemas/Timestamp' }, { type: 'null' }] }
        createdAt: { $ref: '#/components/schemas/Timestamp' }
    DomainResource:
      oneOf:
        - $ref: '#/components/schemas/User'
        - $ref: '#/components/schemas/Ledger'
        - $ref: '#/components/schemas/Category'
        - $ref: '#/components/schemas/Entry'
        - $ref: '#/components/schemas/Debt'
        - $ref: '#/components/schemas/DebtTransaction'
        - $ref: '#/components/schemas/RecurringRule'
        - $ref: '#/components/schemas/RecurringRun'
        - $ref: '#/components/schemas/SalaryProfile'
        - $ref: '#/components/schemas/SalaryRecord'
        - $ref: '#/components/schemas/SavingGoal'
        - $ref: '#/components/schemas/SavingContribution'
        - $ref: '#/components/schemas/Notification'
    ResourceResponse:
      type: object
      required: [success, data, requestId]
      properties:
        success: { const: true }
        data: { $ref: '#/components/schemas/DomainResource' }
        requestId: { type: string }
...
    CategoryListResponse:
      type: object
      required: [success, data, requestId]
      properties:
        success: { const: true }
        data:
          type: object
          required: [items, permissions]
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/Category' }
            permissions:
              type: object
              required: [canCreate, canReorder]
              properties:
                canCreate: { type: boolean }
                canReorder: { type: boolean }
        requestId: { type: string }
...
    StatisticsOverviewResponse:
      type: object
      required: [success, data, requestId]
      properties:
        success: { const: true }
        data: { $ref: '#/components/schemas/StatisticsOverview' }
        requestId: { type: string }
    StatisticsOverview:
      type: object
      additionalProperties: false
      required:
        [ledgerId, ledgerType, month, incomeCent, expenseCent, balanceCent, averageDailyExpenseCent, largestExpenseCent, entryCount]
      properties:
        ledgerId: { type: string, format: uuid }
        ledgerType: { type: string, enum: [PERSONAL, COUPLE] }
        month: { type: string, pattern: '^\d{4}-(0[1-9]|1[0-2])$' }
        incomeCent: { type: integer }
        expenseCent: { type: integer }
        balanceCent: { type: integer }
        averageDailyExpenseCent: { type: integer, minimum: 0 }
        largestExpenseCent: { type: integer, minimum: 0 }
        entryCount: { type: integer, minimum: 0 }
    StatisticsTrendResponse:
      type: object
      required: [success, data, requestId]
      properties:
        success: { const: true }
        data: { $ref: '#/components/schemas/StatisticsTrend' }
        requestId: { type: string }
    StatisticsTrend:
      type: object
      additionalProperties: false
      required: [ledgerId, ledgerType, month, items]
      properties:
        ledgerId: { type: string, format: uuid }
        ledgerType: { type: string, enum: [PERSONAL, COUPLE] }
        month: { type: string, pattern: '^\d{4}-(0[1-9]|1[0-2])$' }
        items:
          type: array
          minItems: 28
          maxItems: 31
          items:
            type: object
            additionalProperties: false
            required: [date, incomeCent, expenseCent]
            properties:
              date: { type: string, format: date }
              incomeCent: { type: integer, minimum: 0 }
              expenseCent: { type: integer, minimum: 0 }
    StatisticsCategoriesResponse:
      type: object
      required: [success, data, requestId]
      properties:
        success: { const: true }
        data: { $ref: '#/components/schemas/StatisticsCategories' }
        requestId: { type: string }
    StatisticsCategories:
      type: object
      additionalProperties: false
      required: [ledgerId, ledgerType, month, type, totalCent, items]
      properties:
        ledgerId: { type: string, format: uuid }
        ledgerType: { type: string, enum: [PERSONAL, COUPLE] }
        month: { type: string, pattern: '^\d{4}-(0[1-9]|1[0-2])$' }
        type: { const: EXPENSE }
        totalCent: { type: integer, minimum: 0 }
        items:
          type: array
          items:
            type: object
            additionalProperties: false
            required: [categoryId, name, icon, color, isEnabled, amountCent, basisPoints, entryCount]
            properties:
              categoryId: { type: string, format: uuid }
              name: { type: string, minLength: 1, maxLength: 50 }
              icon: { type: string, minLength: 1, maxLength: 50 }
              color: { type: string, pattern: '^#[0-9A-F]{6}$' }
              isEnabled: { type: boolean }
              amountCent: { type: integer, minimum: 0 }
              basisPoints: { type: integer, minimum: 0, maximum: 10000 }
              entryCount: { type: integer, minimum: 0 }
    StatisticsMembersResponse:
      type: object
      required: [success, data, requestId]
      properties:
        success: { const: true }
        data: { $ref: '#/components/schemas/StatisticsMembers' }
        requestId: { type: string }
    StatisticsMembers:
      type: object
      additionalProperties: false
      required: [ledgerId, ledgerType, month, totalCent, items]
      properties:
        ledgerId: { type: string, format: uuid }
        ledgerType: { type: string, enum: [PERSONAL, COUPLE] }
        month: { type: string, pattern: '^\d{4}-(0[1-9]|1[0-2])$' }
        totalCent: { type: integer, minimum: 0 }
        items:
          type: array
          items:
            type: object
            additionalProperties: false
            required: [userId, nickname, avatarUrl, memberStatus, isCurrentUser, amountCent, basisPoints, entryCount]
            properties:
              userId: { type: string, format: uuid }
              nickname: { type: string, minLength: 1, maxLength: 100 }
              avatarUrl: { type: [string, 'null'], format: uri }
              memberStatus: { type: string, enum: [ACTIVE, LEFT] }
              isCurrentUser: { type: boolean }
              amountCent: { type: integer, minimum: 0 }
              basisPoints: { type: integer, minimum: 0, maximum: 10000 }
              entryCount: { type: integer, minimum: 0 }
...
    ErrorResponse:
      type: object
      required: [success, code, message, details, requestId]
      properties:
        success: { const: false }
        code: { type: string, pattern: '^[A-Z][A-Z0-9_]+$' }
        message: { type: string }
        details: { type: object, additionalProperties: true }
        requestId: { type: string }
...
    CreateCategoryRequest:
      type: object
      required: [ledgerId, type, name, icon, color, idempotencyKey]
      additionalProperties: false
      properties:
        ledgerId: { type: string, format: uuid }
        type: { $ref: '#/components/schemas/EntryType' }
        name: { type: string, minLength: 1, maxLength: 50 }
        icon: { $ref: '#/components/schemas/CategoryIconKey' }
        color: { $ref: '#/components/schemas/CategoryColor' }
        idempotencyKey: { $ref: '#/components/schemas/IdempotencyKey' }
...